+++ src/settings/MappingProfiles/DuplicateMappingProfileRoute/DuplicateMappingProfileRoute.js
@@ -32,6 +32,7 @@ export const DuplicateMappingProfileRouteComponent = ({
   const newMappingProfileRecord = {
     ...mappingProfileRecord,
     default: false,
+    locked: false,
     name: intl.formatMessage(
       { id: 'ui-data-export.copyOf' },
       { value: mappingProfileRecord.name }