+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Include metadata in edit form of data export mapping profile. Refs UIDEXP-408.
 * Eliminate identifier selection for exports with default profiles. Refs UIDEXP-413.
+* Locking and unlocking mapping profile. Refs UIDEXP-414.
 
 ## [7.0.0](https://github.com/folio-org/ui-data-export/tree/v7.0.0) (2025-03-12)
 [Full Changelog](https://github.com/folio-org/ui-data-export/compare/v6.2.0...v7.0.0)